Refactor asset processing and shader checking
