--- Editor/Scripts/ShaderChecker.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1bff7a1b978c09a4bb5a5de0a9c11685
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
